"9:30am. I am up and I am getting hit with lethargy. Lethargy for anything other than completing the scene. I do not feel like doing any more 'practice'. It is time to move on from that.
9:55am. Yeah, this feeling is rising up throughout my body. The feeling of having had enough. I do not want any hassle at all from art. I'll know right away whether this approach is viable or not.

I feel like quiting with what I've been doing before. I will focus all of my being on this single strike. It won't just land on the target. It will land on it absolutely.

https://mangadex.org/title/c288b108-5162-4065-aa3a-5857ec38c8d9/hikikomari-kyuuketsuki-no-monmon

Let me take this break to catch up with the two chapters of this that are out there and I will get into it. I feel like taking a break so I can roll it around in my mind.

10am. Right now I am just playing with disappointment. 3d is fun. 3d is how I think, I can in fact make art by directly manipulating objects in 3d space. But doing 2d from scratch. Just how am I supposed to int my way to understanding perspective. It would take a shitload of practice.

I am super lazy. I could use 3d as a teacher, but I do not feel like it. I do not feel like working hard on it, what I need to do is simply get it done in the simplest possible manner. 3/5 is more than enough for me. It is fine if I am a bad artist doing good art. Right now I am a good programmer capable of doing absolutely nothing with it. It will be my revenge towards the world for putting me in this situation.

10:45am. Let me start.

Yeah, let me start by blocking out the bed.

11:30am. I have to admit, Boxcutter is quite nice here. Being able to draw boxes directly onto the surface instead of having to make a cube and rescale is really proving its worth.

I've already blocked out 50% of the room. Let me block out the rig and the stool it is sitting on. As well as the desk.

12:30pm. Let me take a break here. I can't get it anymore right than this. It is not exact, and I am not sure where I am going wrong, but it is good enough. I'll do the doorframe and the curtains. After that I'll import the thing into CSP and start painting."