"9:25am. I am up, but I am hesitating a bit. I've been very focused on the subject of how to draw folds. I am not going to go with the plan from yesterday. Rather I am going to lower the size of the pen to the minimum and see where that gets me. I do want to go with the painterly approach, just not all the way with it. The style is very beatiful when it works well. I just need to upgrade my skills with it.
Right now let me read Frieren.

10:05am. Let me start. Let me check out the Moi thread.

10:10am. Oh it seems the boolean difference leaves some planes that I've missed. Let me try it out. After that I'll get back to CSP.

10:15am. Yeah, it works very smoothly. Nice. Let me take the guy for taking the time.

10:25am. I think I am starting to get the feeling. During the night I've considered that it might be better to make many smaller strokes and blur them out than large ones. I really missed something very basic here. This also make it much easier to do the fresnel for the creases. It is a real hell to draw a shadow in a single stroke and then a smaller shadow next to it. But drawing 10 strokes for the shadow and 2-3 strokes for the highlight is not problem.

The most important thing is to get the feeling for it.

10:35am. Let me just clean my monitor a little.

10:45am. It wasn't as dirty as I thought it would be. Let me resume. Right now it is spotless.

Yeah, I am liking this way of doing things. I overdone it a bit on the initial higlights I'll have to dampen them a bit. I thought I did not get much from Inio, but I guess watching him draw those thin beautiful lines did inspire me a little.

12:05pm. Let me take a break here. What I've done is not good by any means, but...it is about as good as somebody just putting down lines without a great model could do. This is just about as one without great understanding could make it.

It is not good, but it has appeal. I haven't made any mistakes, managed to do the folds and the hightlights and made sure to break up the flatness using smaller strokes where there aren't any big folds like in the center. I am not confused as to how to do the highlights like yesterday. I guess I did need to think for an entire day just to decide to lower the pen size.

12:10pm. A mistake I've made compared to the real thing is that the folds are too big and not numerous enough, but that is fine.

This kind of style is not supposed to be realistic by any means. I am just simplifying and immitating reality here. There is only so much I can do with a pen.

12:10pm. Going from size 3 to 1 really feels like it gave me a lot of room to move in.

12:15pm. Now...

https://www.youtube.com/watch?v=lx6B7XRu3nk
Creating Comic Art in Zbrush with Pablo Munoz Gomez | PIXEL PEEPING

I found this last night, but haven't watched it. It seems he is doing 3d models in Zbrush, and then using some kind of comic rendered to get something that looks it was made in 2d.

I feel satisfied now that I've done the bed. I'll watch the above and then get busy putting down flats for the bed and well as do the lines for the rest of the room. Even though I am not following what I said I would do last night directly, not acting robotically is good. Work is best when I am doing it at a leisurely pace. I am going to gradually cover all of the areas that I am still weak in. And one of the most important thing in knowing how to do 2d is knowing how to use small pen sizes properly. I have enough intelligence to at least do that."