"9:10am. I've been thinking what my problem is when it comes to painting that bed. I've realized what my problems are. During the night, I've approached it as a drawing problem instead. What if I simply tried to do line art for this? Then I realized the truth. The creases are essentially noise patterns, so it is no wonder that I can't store them in memory and comprehend them. I knew that fundamentally the way I am approaching them is broken, just putting down some strokes and then bluring them out.
It simply doesn't make sense to me. And because the goals are so ambiguous I find it impossible to draw. The end result is as if I were just tossing mud at the screen. Yesterday when I finished I was happy for 10m until I reviewed and came to my senses.

Forget painting, and trying to shade the bed directly. I already know that is very hard. Instead what would be something that I could draw? As weak as my 2d skills are, I should be able to draw primitive shapes.

Since the bed is a 3d box, imagine if I scattered a bunch of cylinders over it and tried drawing that. Imagine trying to draw curved surfaces like the ones I've made in Moi.

Could I draw this curved surface, could I draw that curved surface? What is the problem, and what is my limit?

9:20am. If I focus on composing primitives while 2d drawing, I should be able to make progress. But instead I started off with shading, did not at all feel out the 3d shape of what I was painting and got completely lost.

9:25am. Somewhat ironically, this kind of approach of focusing on primitives would make 2d and 3d art quite similar, but that is good. That is how I would know I am on the right path when it comes to 2d. 2d should not be completely different from 3d.

9:30am. The painterly approach of trying to optimize the shades would work if I had a really good reference, but changing brightness so they match is fundamentally different from stroking out the iso-curves of various surfaces. I can't use the former approach to feel out the 3d space. Unless I have it to begin with I am going to get lost.

I lost the moment I flat shaded the planes of the bed. That is the part where I went wrong.

9:35pm. While programming I often write code and then think about how it fits in. There is only so much you can do by trying to think everything through from the beginning. The hand takes the office of the mind.

I absolutely have to make it so that this applies to 2d art as well. It certainly applies to 3d art where I am constantly moving things around.

9:35am. If I could learn how to manipulate primitives directly in 2d space, and learn how to draw curved surfaces, then I could move from the bottom of 2/5 to somewhere decent. The first step towards acquiring skill is to set sensible goals. There is absolutely no reason to be distressed over having bad memory and such.

If I can learn how to draw curved surfaces, then I can learn to compose them in a scene to get a good looking image. At point I can do the coloring and the shading. If I had the creases in the reference, I would not have failed at painting them.

9:40am. This all goes back to what I studied last year. Remember those Drawing Database vids where he spent all his time drawing primitives? I want to say I forgot the fundamentals, but I never really learned them to begin with. I sort of didn't see the point since I had 3d. Why practice drawing boxes directly when I could just import them?

And indeed, that is a good question - what was I even trying to do in the past couple of days?

Apparently, not use my brain, go somewhere I should not be, and get my ass kicked. I should learn from this experience. Forget trying to immitate those experienced painters who can do what they want from imagination directly. The loose approach is faster for them, but it won't be for me.

Tracing out that couch which I am still using as the wallpaper for my desktop only served to retard my progress.

9:55am. It is not that something like this has never happened in programming. I start writing code and then I find that I can't keep track of what I am doing in my head. The plan turns out to be convoluted and I become lost as to what it is I am doing.

10am. Right now I am thinking about my wins and losses.

I am thinking about the documentary of Inio Asano. I hadn't really thought about taking an image and leveling it out before putting in my own drawings.

Suppose I wanted to take such an approach.

If I just took a 3d photo as it is, then I'd have to match the level of detail on my own model. but if I grayscaled it and leveled it out, I would have to do a lot less work to make the model integrate with the background.

I mean, if I was doing it grayscale, I would not need to color or texture my model. I could just take the same approach as Pablo or use the Mangaka plugin to integrate it into the scene.

10:05am. His approach of using photographs would be viable whether I am using 2d or 3d.

So in fact, given that it saves me from directly modeling a lot of things, it would be a huge gain. I can always go into a grayscale image and paint it myself so it matches the simplified texturing of the model.

10:10am. I know deep down that if I want to get good results, it is not realistic of me to model literally everything from scratch.

I am going to have to learn to draw if only so I can retouch leveled-out photos and some painting so I can color those images back in.

10:20am. This brings me to the present situation. What is my plan?

I do not need to start grinding 2d. Doing more 3d modeling should be more than good enough. I mean, Moi and Zbrush will make me better at 2d drawing, than 2d ever will.

So I should really level up my modeling skills as much as possible.

For painting, I will check out the NPR stuff in Blender. No doubt I will be using a combination of that and hand painting in CSP. It won't be fancy at all.

10:40am. That huge and intricate spaceship Inio did? Since it is grayscale, he could have just painted that intricate machinery using black lines. Since the depth is leveled out, I'd naturally imagine it would be some huge and intricate 3d modeling job, but it could not be further from the truth.

This is real magic. It is natural to imagine great artists being very skilled, but no amount of skills will give those kinds of results.

10:50pm. Let me set the course to download.

10:55am. Yeah, the plan is simple. In order to be able to take a few steps forward, I am going to have to take a step back first. I spent time learning texturing, but that was going in the wrong direction. Somewhat ironically, the Designer stuff was in the right direction as it directly relates to compositing.

I am going to have to go to grayscale and then back to color. That is what I should aim for. It will make my art much easier to integrate with photo references.

For painting my models I am going to have to look into NPR. Maybe I won't need that and will opt to do my paint over directly in CSP. But it does not matter. Either way it is not a big deal.

11am. I can leave investigating NPR for later.

My current approach will place modeling front and center. This is just the way I like it. I did not want to do the rest all that much. I found texturing interesting, but only to a degree. I knew that I couldn't spare the time for it.

I won't get obsessed about NPR at this point. I'll just have faith that I'll deal with it when the time comes.

For now what I should focus is on finishing the room. I should imagine my intention is to draw it and then paint it as if I were doing 2d. Forget NPR, and forget everything else.

11:05am. Let me have breakfast here first. The night was a bit strained for me as I was obsessed about the painting failure. Right now I still feel groggy."