"7pm. Let me do some more modeling. I am feeling up for it now.
7:25pm. I am running into trouble with the basic boxes. I could leave this for the later stages, but let me just ask.

7:35pm. https://moi3d.com/forum/discussion.php?webtag=MOI&msg=10705.1

Here is the Moi issue. Let me make the other box. This is not that important to me, so I might as well leave it for later.

8:40pm. Modeled the USB stick. Not bad.

Got some experience with using arcs. I've been unfairly neglecting them, I should have been using it instead of freeform in places. The point as well.

8:45pm. https://danbooru.donmai.us/posts/5367063?q=fuzichoco

Fuzichoco (Kenja No Deshi's novel illustrator) is very good. When I say that old masters have nothing on the new ones, people like him are exactly what I mean. People wonder where the von Neuman have gone off to, but there is need to ask the same thing about the artists.

This is what rank 5 in 2d would be for me.

I imagined if I got better in 3d I would be able to make these kinds of scenes quickly, but I am going to have to gain some skill in 2d if at least to comparmentalize my 3d work.

8:50pm. This kind of modeling I am doing here is really great for internalizing various shapes. Sure I know what an USB is, but this kind of exercise really forces me to study the form in ways that aren't superficial. I am starting to think that there is a beauty to hard surface models. Had I made the choice not to go down this path, I never would have realized this."