"9:30am. Am just chilling. Let me do it for a while longer and then I will start. Rosen Garten is out.
10:15am. Let me start. First thing's first, let me check out the Moi thread.

https://moi3d.com/forum/discussion.php?webtag=MOI&msg=10705.1

10:25am. From the way I've been using the pen I am starting to feel some hand muscle issues. I need to relax it as much as possible. And try to use the elbow and shoulder more. Maybe it would not be bad to take a day off from this.

Nevermind that. I am only a day from being done. Let me investigate those boxes.

10:40am. Let me try opening the gaps myself as an exercise. After that I'll get started on modeling the rest of the items.

///

10705.6 In reply to 10705.1
Hi Markog, in the attached version I've opened up that little bit of space in between the flaps by copying edge lines and using Edit > Trim and also cutting out the margin in the corners as well.

With this setup you should now be able to shell the whole thing ok, set Shell thickness = 0.05 and set Direction: Flip.

The spacing helps to keep the flaps from running in to each other.

///

I am not really sure how it would be possible to do this efficiently using trim. But drawing rectangles and using boolean diff works quite well. Let me pause this for a bit. I want to check out the usb just a tad. I remembered an idea from last night.

11:10am. It feels like too much of my time is lost due to posting stuff. At any rate I've tried out my idea. I misremembered - 3 pt arc do not have the eliptical feature. They are already naturally that way.

But conics are really good, that are exactly what I need for what I had in mind. Let me rebuild the body of the USB...no forget it. I put a chamfer on the rest. It is not worth it for me to mess with this.

11:20am. Let me deal with the other box. Focus me. Actually lret me repeat what I've done on the first box. The gap needs to be a bit wider.

11:35am. Ah damn, sliding along a rectangle can sometimes not do a clean job.

12pm. Shit, now I am losing time on this. Why is the gap not wide enough. Doing this kind of thing in Zbrush would allow me to avoid these kinds of issues. I should just leave the box for the later stage instead of bashing my head with Moi.

12:20pm. I finally did the box. It took me far to long to realize that Flip means the flip normal direction. Anyway, what is next?

12:30pm. Ah shit, I just realized I could have done those stripes on the router using inset. What I did was shell a copy of it, take the intersections, and the diff it with the original which is a lot more long winded. Instead I should have used trim to cut in the new faces and inset those isntead.

12:35pm. I am distracted. I need to get started. But actually this is a good place to stop for the morning. Let me have breakfast here."