">>899672 Done for today. It is really tedious to just watch tutorials, but tomorrow I should get to the NPR and Photoshoping parts. Today, I've been boring myself watching him model and the only thought on my mind was that could do all of this just as well as he could. Right now my problem is that I am split on whether I should dive deeper into shading arcana with Malt, or whether I should just extract the freestyle edges and the rendered scene and play around with gradient maps in CSP. I'll know soon enough where I stand. Malt strikes me as particularly powerful, but I know that messing with shaders is not the most direct way to go about this. I know that if I really dove into it, it could eat up a month or two of my time. If I was doing animation it would be worth it, but right now I have a lot leeway, too much in fact. I am really opening up to how much 2d stylization makes things easier - models can have bad topology, I do not need to bevel, UV unwrap or texture. Even the geometry does not have to be good, as long as it looks good past the NPR filters, everything is fine. If I want to make it, I need to focus on taking advantage of all of this. The question is merely, what is the most efficient way of doing that?
When it comes to rendering, my hunch is that just letting Cycles or Luxcore will be hard to beat, as they don't need fiddling around. I won't have to let them run for hours in order to clean up an image like I have in the past. Since I will be crunching the midtones, the benefit of that should be minimal. A few minutes followed by denoising should be fine to get me the correct lighting.

I am sure that the closer I can get to a workflow that focuses purely on modeling, the closer I will be to winning."