"9:10pm. I am thinking how I should tackle the problem of making my own NPR filter. To begin with, what I should do is test whether the idea is viable at all. I think the ideal way to do it would be to try it out in Substance Designer. I should be able to use the shape splatter with the original image as a density mask to distribute various kinds of strokes.
The very first challenge is to extract the normal information from the renderer, but that should be manageable. The second will be to figure out how to use the shape splatterer to make it do what I want. Assuming I could get a decent result from this, I can think about future steps. Rather than work on a Blender addon that does this, I should refine the technique in Designer first and see how far I can get with it. If it turns out to work well, I'll have an essential element of my own style.

I should be able to implement the above, but maybe it will turn out to work poorly. Who knows. The idea is quite simple though, I am not sure why I haven't seen it being used.

My understanding of painting is that it is about using values as probability density for strokes. There is also the understanding of directionality to consider, but doing them orthogonal to the normals is the solution. Note that I can't do this later thing using just pixel values, so this is why PS does not have such a filter.

The idea I have here cuts directly to the meat of the issue of what painting really is, so I will be quite disappointed if it turns out to work poorly. It is worth exploring.

If it fails, I suppose I could just become a B&W guy, and only put in special effort when it comes to doing covers. I'd like to avoid letting it come to that if it is at all possible."