"1:45pm. Let me resume the course. I'll get it out of the way and then I'll be done with this drudgery.
The main problem with the course is not that it is boring. It is that Pablo's style is not what I am aiming for. I'd prefer a clean, illustrated anime style. If I could do it like Fuzichoco for example, that would be perfect. But if I had to settle for a rougher style, I'd still not want to do it like Pablo because it is too asexual. The colors are too flat, and the linework is too realistic. It is not quite what I am looking for.

But let me get all the tips that I can. I only have 3h of the course left so let me just go through it.

2:05pm. Enough, I am just wasting my time. Let me start skipping.

2:10pm. Let me move on to the last part. I am really tired of this as I already decided what I should be doing.

2:15pm. This line refinement is interesting.

He brought in differnet kinds of linework, and is using masks to blend between them. He starts out with a thin line layer and is bringing in the thicker ones where detaul would be needed. I forgot exactly where the thicker lines are supposed to come from, maybe the render with the cavity outline. Zbrush is really bizzare with how it only has 12 filter slots instead of having a layer system.

2:25pm. Ok, this caught my attention a bit, but let me not watch all the work mode stuff. Let me skip.

2:35pm. Let me pause at 33m.

3:10pm. Let me resume. It is time to finish the course. I sure had enough of it.

3:30pm. https://youtu.be/X8YkWdhty7I?t=563

Let me watch this just for a bit. I need to think about it.

The thing here where he splits HSV and the posterizes them directly would be quite difficult to do using gradient maps in CSP. Maybe I really need a proper compositor for this. Either Designer or maybe Nuke.

https://youtu.be/uVRdep73hCQ
Substance Painter Stylized Hand Painted Texturing - Tutorial for Beginners

I'll watch this later.

https://docs.blender.org/manual/en/latest/render/shader_nodes/shader/toon.html

Cycles specifically has a toon BSDF.

Let me get back to the course, I'll soon be done with it.

https://www.foundry.com/products/modo/addons/non-photorealistic-rendering?_ga=2.223160348.666735247.1547165221-360079755.1545431136

Focus me, let me just finish the course.

4:10pm. Enough I am done. I could have skipped this course in its entirety and it would not have impacted me too much. But it did get me thinking of how removing details can make it easier on me. Even something like decimation which I've thought to be useless played a role in optimizing the assets.

https://learn.foundry.com/course/6060/view/introduction-to-nuke-from-actionvfx-learning-the-basics

Let me watch this just for a bit. I want to understand what Nuke is about.

https://youtu.be/u6UkJRuN6j0?t=81

Hmmm, is that how it is?

4:20pm. https://youtu.be/hQ9nEYCwYDQ?t=41
> It is 40h long.

Damn son.

Forget it, this is too advanced. This is some weird kind of programming.

Let me just gather my thoughts. I'll write down all the avenues I need to explore. Before that, let me watch the above.

https://youtu.be/uVRdep73hCQ
Substance Painter Stylized Hand Painted Texturing - Tutorial for Beginners

I'll watch this. Then I will watch the illustration shader vid. Maybe I'll even buy it just so I can see what it will do.

* Procedural hatching & painting via shape splattering
    * Need to extract depth, normals and image info
    * Also want to check out the illustration shader for Blender
    * Easiest to try out in Substance Designer
* Line Art
    * Need to check out Blender*s Freestyle
    * Should be a good substitute for small beveling
    * Could also be useful for defining inner cavities
* Toon Shading
    * Colored:
        * Can be done at render time using shaders
        * Can be done and compositing time using gradient maps and posterization in either Blender compositor or Clip Studio Paint.
        * In CSP I could also do some hand blurring
    * Grayscale:
        * Need to try out the Mangaka plugin
        * Can be done using PS filters and some manual redrawing
        * Has the advantage of easily being integrated with real life refs. Is incompatible with procedural hatching which requires the scene to be all 3d.

Let me watch those two tutorials and then I will elaborate on the above points.

https://youtu.be/uVRdep73hCQ?t=6

This painterly look is quite nice. I thought that Painter could be used like this.

4:50pm. https://www.youtube.com/watch?v=rXdeJ_sigf4
Blender Illustration Shader - Eevee Live Lighting Tutorial

Let me watch this. I started it a few days ago, but quickly lost interest.

https://youtu.be/rXdeJ_sigf4?t=206

If I had to implement this myself - remember the leaf tutorial by Creative Shrimp where he instanced it on Voronoi Texture cluster centers? Yeah, that is what I would have done.

https://youtu.be/rXdeJ_sigf4?t=980

I do not get it. Why are geo nodes an inputs to the shader.

5:10pm. Nevermind this, I skipped most of it.

I really had enough of these tutorials which are just like a encyclopedia of different features these programs have.

I give up. There is no point looking for anything else. Though I am not sure how he implemented the illustration shader, the concept of distributing strokes over an object is not hard to grasp.

In Zbrush I thought Pablo might be doing something really smark to get the object ids, but he is just polypainting and polyframing and then using the equivalent of the magic wand in PS to create the masks.

5:15pm. Sigh. Ok. Let me try out the Mangaka plugin. I want to see how my room looks in grayscale. It might turn out that I am not getting anything I couldn't get in CSP.

https://youtu.be/DhydDgiZV1w
MANGAKA!

It is acting strange. How do I use this thing?

I've tried it out on a sphere. It is not that big of a deal. Let me move on.

https://docs.blender.org/manual/en/latest/render/shader_nodes/input/camera_data.html

Hmmmm...how do I use this to get the surface normals relative to the camera?

https://blender.stackexchange.com/questions/69317/cycle-normal-pass-in-camera-space

This might be an answer. Is there an easier way?

Also how would I output this as a separate pass instead of putting it on an emission shader.

https://blenderartists.org/t/screen-or-camera-space-normals/702009/8
> You don’t need to modify any shaders or materials. What you need is two additional renderlayers with two materials that overrride all scene materials. One produces the world space normals, other camera view vectors. And from these two layers you should be able to produce the camera space normals in compositor.

How would this work?

6:10pm. https://blender.stackexchange.com/questions/157271/render-normals-in-camera-space

How annoying this is. Just what am I supposed to do with world space normals in the compositor? Well, I'll figure it out with some effort.

7pm. Done with lunch. I see that shaders have an AOV Output node. I'll figure out how to make use of it tomorrow.

7:10pm. In the end I didn't get a real lead on how I want to do things, but I'll able to come to a conclussion after I try out procedural hatching in Substance Designer. Even if it fails, I'll be able to make a cell shaded style work so it is not huge of a deal, but I do not want to give up this easy chance to stand out.

Let me do a /wip/ post and then I ma done. I see an AyaTri chapter is out."