"10:10am. This tension really reminds me of programing. Let me start. I'll try out the hatching thing right now. I do not even need the AOVs, let me just set up the rest of the scene so it fits the came at zero rotation. Not the room scene, but a simple sphere scene.
10:35am. This is one thing that has been irking me. How do I run the compositor in order to get the output?

10:55am. Sigh, what is going on with depth? It does not get saved properly. Nevermind, let me just import these files in Designer.

11am. Hmmm, the stupid thing does not allow me to set arbitrary resolution. Let me render it in 2048 x 2048. This is just an experiment right now.

Let me take a break. By now I've forgotten how to use substance designer.

11:35am. Damn it, this is so annoying. How do I normalize a vector in the compositor? Let me try doing that in Designer's pixel processor instead.

https://www.youtube.com/watch?v=lFfgC0tFiAU
Normal map Rotate part 01 (Allegorithmic Substance Designer 5)

Let me watch this for a bit. I forgot how to get the input variable in the pixel processor at this point.

12:20pm. Fuck. No, I went in the wrong direction. To begin with, all the shape splatterers generate a tile and then randomize them. If I want to use the original image as a mask, at best I get a result like the original image, but with parts of it missing. It looks exactly like the original image being put through a noisy threeshold. It looks like shit.

Furthermore, these substance nodes aren't a good fit for what I want to do because they generate evenly spaced inputs while I want them to randomly scatter them. Furthermore, it is a massive pain in the ass to use the pixel processor of Designer.

I've been thinking about this in the wrong direction. I should be using geo nodes for making these kinds of experiments. That it is in 3d does not matter. I'll put on the isometric view, and use a flat matcap.

Maybe Houdini would be better for this, but let me try it in Blender first.

I feel so uncomfortable right now, like I am on tilt. The tension was so high tonight that I could not get my usual sleep.

12:35pm. Let me take a break here. This is so annoying. It felt like a good idea at first, but I should have seen the thing here today coming. Maybe that Covid really did lower my IQ. Or is my head so filled up at this point that I can't think clearly anymore?

Whatever it is, I have to keep going. I am going to try to make this work either in Houdini or Blender. I'll start with Blender. I'll throw down 100k points, and give them directionality. I'll try loading the files I made as textures and then passing them as density and rotation into the geo node tree. I should be able to use vector to euler. That will be the simplest solution. Designer is such a pain in the ass. If that ends up working, I'll have a basis for everything else.

Maybe I am tense because I want this to fail so I can just get on with it. Common sense is telling me that if this was such a good idea, other people would have done it."