"8:15pm. Oh right, let me also do a little review of Moi 3d. It deserves it. I'll post it tomorrow when the thread is up.
///

Also, let me do a review of Moi 3d. It is a NURBS modeling program. Compared to Blender which deals with polys, it is a lot more restricted, but what it can do, it does much better than Blender even with all the addons paid for.

Good:

* Excellent design. It is a very simple program and it is possible to learn the entirety of the program in half a week compared to months it would take for Blender.
* Optimized for working with pen tablets. That is how I've used it so far and it is very ergonomic.
* When I had trouble and needed help, the author was very prompt in giving me advice how to deal with some corner cases on the forums.
* Once you learn how to model in it, I've found it to be a lot more enjoyable than in Blender. It is easy to get precision with it, and the tools are so on point.
* It can import subdiv models. Its export capabilities are quite good, better than Rhino's.
* Long 3 month trial that I've yet to exhaust.

Ugly:

* To get full use out of it, you need to set the hotkeys yourself. Unlike Blender it does not come with sensible defaults already set.

Bad:

* NURBs modeling can run into difficulties with weird corner cases. When dealing with intersections boolean operations can mysteriously fail. As beginner having to deal with that is very frustrating and the program gives you no indication what is wrong.

I won't put this as a point against it, but I should note that NURBs modeling can't do deformations as well as poly modeling can, so a good workflow would be to use NURBS for hard surface and sketching, subdiv modeling for simple organic models, and sculpting for complex organic modeling and things like adding folds to beds and making blankets. That last thing would be quite difficult to do in Moi. I'd rate Moi 4/5. If it wasn't for the mysterious NURBs corner cases that cause it to fail, it would be a perfectly designed modeling program in my view. As it is, it is merely very good. I wanted to take the time to review it because it is worth it. In Blender I have a bunch of addons like Hops, Boxcutter and MESHmachine, and I really don't need them because working with Moi is much better.

///

Why do I model at all? At this point it does feel like all art is some form of sketching. Good artists can imply more with less lines. No artist can capture reality. I must not lose sight of that."