"9am. Let me chill a bit and I will start.
9:25am. Enough of the Rance thread. I'll leave Clyce and Her Majesty's Swarm for later. It is time to do some sculpting. Ah, let me post the reviews up first.

https://www.reddit.com/r/ProgrammingLanguages/comments/v2sevf/comment/iavya3n/?utm_source=share&utm_medium=web2x&context=3

Here it is. Let me post it in the Twitter thread.

9:40am. Let me get started now.

10:20am. I've seen enough. Let me remesh it and I will use the snake hook brush to get the job done.

10:25am. Damn it, these brushes are so hard to work with. They have huge lag.

Hmmm, how do you switch to another object without having to go through object mode? I think it was D or Shift+D but it is not working for me right now. Remember when I did huge line by accident while working on the base mesh? It is possible to do it.

https://www.reddit.com/r/blenderhelp/comments/mfs7ao/how_do_i_switch_objects_in_sculpt_mode/

https://blender.stackexchange.com/questions/183121/how-to-switch-object-in-sculpt-mode

Let me try Alt + Q. Yep that is it. Nice. Let me remesh the pillow.

10:55am. i might have gone overboard on the bed creases.

Ok, that is done. It looks good, so that is all that matters. For the blankets, I'll just make use of the clothing sims. First let me make the mattress a colider.

Huh, does the collision modifier now work for both rigid body, and soft body and cloth sims? Nice.

I thought it was awkward how rigid body and soft body sims had separate ones.

11:25am. Had to take a break for a bit. Let me try out the clothing sim. Admittedly, I hate doing these, but I should get used to it.

11:30am. Doing sims is so annoying. Let me load another file. I can't tell what is going on at all due to the huge lag.

11:40am. Let me do the chores here.

12pm. Let me resume. I've opened a separate file and tried the cloth sim there. It seems to work fine. I only get trouble once I try messing with remesh modifiers. Let me go back to the original file. Let me replicate what I just did here. Then I will try bending the blanket.

12:15pm. I've forgotten just how much cloth sims suck. They suck really, really bad. Sculpting the folds by hand is totally a viable approach. I complained a bit about the cloth brush, but that is nothing compared to how much the cloth sim sucks.

Let me try sculpting the blanket.

12:15pm. I won't use the 0.005 voxel resolution. It is causing the cloth brush sim to take too long. Instead let go for something like 0.02. It does not look as good, but I'll have to make do. Trying to do it in Zbrush is also an option.

12:20pm. The Enable Collision option in the cloth brush is so slow as to be useless. I'll have to disable it.

12:25pm. Let me do more chores.

12:35pm. I am back. I am really struggling how to do this blanket. I need to take a proper sculpting approach to it. Let me mask out the affected pieces, and cut them. Then I will drag them into place. Trying to use the cloth brush with drag or snake hook is not working for me. I should focus only blocking out the blankets and only use the brushes to give the appearance of the folds being there. With cloth sims I will be here all day at this rate. They are so hard to use. The same goes for the cloth brush.

12:40pm. As annoying as it would be to draw the folds, maybe for a person good at it, it would be easier. I keep seeing that guy in /ic/ posting that drawing is easier and I can't help but feel there is a dash of truth to it. Though I think that style transfer would work better with ray traced 3d scenes, it is not like it would not work with drawings.

Also it is not like I could not draw in 3d, I just need to account for the line work from the start.

And it is not fair to compare drawing to cloth sims. I should compare drawing to reg sculpting. I should try that approach. Forget chara modeling. Just do these blankets.

Let me first have breakfast and then I am going to home in on the problem."