"9:15am. Discovered DALL-E 2 yesterday.
https://youtu.be/U1cF9QCu1rQ
How This A.I. Draws Anything You Describe [DALL-E 2]

Its capabilities here are most impressive, but the way OpenAI gives access to it might as well make it vaporware. I am not sure how big the model is, but it is good bet that I wouldn't be able to even run it on my GPU. It gives me a target to aim for. If I am going to win through art, it only makes sense to get resources needed to train something like it through art.

Having such a system would be a huge boost to my artistic ability. It would not be rank 4 or 5, but rank 6. What it does is beyond human ability and clearly superhuman. It is a pity I can't use it. For the time being I'll go with style transfer.

https://www.youtube.com/watch?v=n9ZNGVvMOSQ
5 Cartoon Tips You Really Need to Know (Blender)

https://youtu.be/5RQLWFxAN2Y
Blender Glass Shader for Toon Shading

Let me take a look at these links.

9:25am. I'll take it easy today. Let me watch these cartoon tutorials, then I'll watch some NN style transfer tutorials. I am not sure whether I want to get started trying to make that work on my own. I'll just do a bit of studying on it, but stick with the DDG for now. It is not worth the hassle. If it was the usual me, I'd sink a month or two into making it work on my own, but I am burnt from that.

https://youtu.be/7wpisn8t3us
4 Options for 2D Lines || Blender 2D

This caught my eye. Let me take a break here first though.

10:05am. Mato chapter and then I will get on with it.

10:25am. https://youtu.be/7wpisn8t3us
4 Options for 2D Lines || Blender 2D

Let me watch this for a bit. Then I'll move on to style transfer vids.

https://youtu.be/7wpisn8t3us?t=890

Come to think of it, what Pablo did with his comic matcap is in fact eqivalent to this, at least as far as getting that outer line is concerned.

10:50am. Ok, enough of that. Yesterday's success made me more hopeful about the future, and as a result I've started relaxing mentally. I need to pick up the slack.  While it true that I might not need PB renderers like Luxcore, I still like the soft lighting as input. The best solution to my current predicament is to get better hardware. But even so, Malt and Eevee should work great with style transfer.

So let me study the later. I am going to need some shader magic to do my thing, but not too much of it. I should focus on style transfer for now.

https://www.youtube.com/results?search_query=neural+net+style+transfer

Let me see what Youtubers have to say. Also I want to look into the original deep dream. I forgot what kind of objective it had that caused it to generate all those dog faces everywhere. I want to make sure it is not regular style transfer.

https://youtu.be/gzQxnZDePko
Neural Style Transfer || A Neural Algorithm of Artistic Style

Let me start with this.

https://youtu.be/gzQxnZDePko?t=256

It seems he shows how to implement it.

https://youtu.be/gzQxnZDePko?t=370

20m params. Assuming they are f32, that means the net should be 80mb. This should be doable on my system.

https://youtu.be/gzQxnZDePko?t=474

The way he calculated the gram matrix is different from how she explained it would be calculated in the other video. On the video, she did not account for the reduction. As expected, code is what you need to understand all of this.

https://youtu.be/gzQxnZDePko?t=597

Style weight, content weight. Here they are.

11:30am. I am guessing that for whatever model they are using the Thin Style ommits the lower layers. That is the only thing that makes sense for what it could be.

https://youtu.be/gzQxnZDePko?t=750

Here he mentions some extra techniques that perform extraordinary style transfer.

https://youtu.be/eeeai1OrnDI
Pix2Pix || Paired Image to Image Translation || Developers Hutt

What is image translation? That is a bit like style transfer, right?

Let me take a short break and then I'll watch more videos. I thought I would leave this alone, but this would be really easy to implement on my own. The programming effort would be no more than a day. Of course selecting the right model and playing around would make it take longer, but I can pay that expense of time.

11:45am. Let me resume. I am going to dive into this.

https://youtu.be/eeeai1OrnDI?t=168

Oh this is an interesting idea. This should act to stabilize the GAN.

12pm. Nevermind this. This is not of interest to me. And the image resolution it too low. For the kind of magic DALL-E 2 does I'd need a lot better hardware.

https://youtu.be/1-26FDZM-Zg
The best style transfer?

Let me watch this.

https://youtu.be/1-26FDZM-Zg?t=25

Yeah, it would be a good idea to get one of the newer papers and implement that.

https://youtu.be/1-26FDZM-Zg?t=105

Oh wow take a look at that drawing of a basketball player. One thing that bothers me about the DDG is that in the image of a rig, it always had lines drawn over it no matter which style I used. I get the sense that this should not happen. Maybe different methods would work better.

https://youtu.be/1-26FDZM-Zg?t=187
> Generating 1k outputs requires 12gb of memory.

Seriously? Something like this I'd need to adapt so it does not go over 3.5k.

Thankfully unlike with poker game sequences, doing this with images will be much easier.

https://youtu.be/1-26FDZM-Zg?t=228

Actually this is weird. Why does 512x512 require 6gb, while an image that is 4x larger require 12gb? This scaling does not make sense.

Also it does seem like the runtimes take a while.

12:30pm. https://youtu.be/1-26FDZM-Zg?t=469

Had to do a bit of chores. Let me resume. The cool thing comapred to doing this instead of poker is that I do not need the latest and greatest. The old thing will do fine here and I can move to newer methods once I get the hardware for them. With poker, unless I can train a very good agent, I have nothing of value. I'll read the paper just to learn the motivation behind it and the problems with older methods.

https://youtu.be/1-26FDZM-Zg?t=561

The transfer to a drawing on the bottom left is really good. Is there a point to using the regular conversion to grayscale anymore. I doubt it.

https://youtu.be/1-26FDZM-Zg?t=682

This is nice. 1,000 out of 1,000 times you'd imagine this was drawn by hand. I'd never have though it possible. Inio Asano can stop redrawing pictures by hand.

This really opens up a lot of possibilities for me.

https://youtu.be/1-26FDZM-Zg?t=773

Let me pause here so I can eat."